Renamed new classes and tests to follow convention.
