+++ lib/SVG/Graph/DataPoint.rb
